Rename plugin asset output display
