+++ lib/config-generator.js
@@ -29,7 +29,7 @@ const cleanPluginUtil = require('./plugins/clean');
 const commonChunksPluginUtil = require('./plugins/common-chunks');
 const productionPluginUtil = require('./plugins/production');
 const friendlyErrorPluginUtil = require('./plugins/friendly-errors');
-const devServerPluginUtil = require('./plugins/dev-server');
+const assetOutputDisplay = require('./plugins/asset-output-display');
 
 class ConfigGenerator {
     /**
@@ -231,7 +231,7 @@ class ConfigGenerator {
 
         if (!this.webpackConfig.useDevServer()) {
             plugins = plugins.concat(
-                devServerPluginUtil.getPlugins(this.webpackConfig)
+                assetOutputDisplay.getPlugins(this.webpackConfig)
             );
         }
