+++ lib/plugins/asset-output-display.js
